+++ app/src/main/java/com/wear/pin/core/network/JsonProvider.kt
@@ -1,0 +1,12 @@
+package com.wear.pin.core.network
+
+import kotlinx.serialization.json.Json
+
+object JsonProvider {
+    val networkJson =
+        Json {
+            ignoreUnknownKeys = true
+            explicitNulls = false
+            isLenient = true
+        }
+}