+++ app/src/main/java/com/wear/pin/core/network/OkHttpProvider.kt
@@ -1,0 +1,21 @@
+package com.wear.pin.core.network
+
+import com.wear.pin.BuildConfig
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+
+object OkHttpProvider {
+    fun provideOkHttpClient(): OkHttpClient {
+        val builder = OkHttpClient.Builder()
+
+        if (BuildConfig.DEBUG) {
+            val loggingInterceptor =
+                HttpLoggingInterceptor().apply {
+                    level = HttpLoggingInterceptor.Level.BODY
+                }
+            builder.addInterceptor(loggingInterceptor)
+        }
+
+        return builder.build()
+    }
+}