+++ app/src/main/java/com/wear/pin/core/network/RetrofitProvider.kt
@@ -1,0 +1,20 @@
+package com.wear.pin.core.network
+
+import okhttp3.MediaType.Companion.toMediaType
+import okhttp3.OkHttpClient
+import retrofit2.Retrofit
+import retrofit2.converter.kotlinx.serialization.asConverterFactory
+
+object RetrofitProvider {
+    private const val BASE_URL = "https://api.pinterest.com/"
+
+    fun provideRetrofit(okHttpClient: OkHttpClient): Retrofit {
+        val contentType = "application/json".toMediaType()
+        return Retrofit
+            .Builder()
+            .baseUrl(BASE_URL)
+            .client(okHttpClient)
+            .addConverterFactory(JsonProvider.networkJson.asConverterFactory(contentType))
+            .build()
+    }
+}